chore: FinancialRecord with values from excel

+++ src/main/java/com/andrerog/core/FinancialRecord.java
@@ -2,5 +2,5 @@ package com.andrerog.core;
 
 import java.util.Date;
 
-public record FinancialRecord(Date date, float value, String description, float finalBalance) {
+public record FinancialRecord(Date date, double value, String description, double finalBalance) {
 }
